rename status to foghorn
