+++ charts/lighthouse/templates/_helpers.tpl
@@ -21,7 +21,7 @@ We truncate at 63 chars because some Kubernetes name fields are limited to this 
 {{- printf "%s-%s" .Chart.Name $name | trunc 63 | trimSuffix "-" -}}
 {{- end -}}
 
-{{- define "status.name" -}}
-{{- $name := default "status" .Values.status.nameOverride -}}
+{{- define "foghorn.name" -}}
+{{- $name := default "foghorn" .Values.foghorn.nameOverride -}}
 {{- printf "%s-%s" .Chart.Name $name | trunc 63 | trimSuffix "-" -}}
 {{- end -}}
