+++ charts/lighthouse/templates/foghorn-crb.yaml
@@ -1,12 +1,12 @@
 kind: ClusterRoleBinding
 apiVersion: rbac.authorization.k8s.io/v1
 metadata:
-  name: {{ template "status.name" . }}-{{ .Release.Namespace }}
+  name: {{ template "foghorn.name" . }}-{{ .Release.Namespace }}
 roleRef:
   apiGroup: rbac.authorization.k8s.io
   kind: ClusterRole
-  name: {{ template "status.name" . }}
+  name: {{ template "foghorn.name" . }}
 subjects:
 - kind: ServiceAccount
-  name: {{ template "status.name" . }}
+  name: {{ template "foghorn.name" . }}
   namespace: {{ .Release.Namespace }}
