+++ charts/lighthouse/templates/foghorn-rb.yaml
@@ -1,11 +1,11 @@
 kind: RoleBinding
 apiVersion: rbac.authorization.k8s.io/v1
 metadata:
-  name: {{ template "status.name" . }}
+  name: {{ template "foghorn.name" . }}
 roleRef:
   apiGroup: rbac.authorization.k8s.io
   kind: Role
-  name: {{ template "status.name" . }}
+  name: {{ template "foghorn.name" . }}
 subjects:
 - kind: ServiceAccount
-  name: {{ template "status.name" . }}
+  name: {{ template "foghorn.name" . }}
