+++ charts/lighthouse/templates/foghorn-role.yaml
@@ -1,7 +1,7 @@
-kind: ClusterRole
+kind: Role
 apiVersion: rbac.authorization.k8s.io/v1
 metadata:
-  name: {{ template "status.name" . }}
+  name: {{ template "foghorn.name" . }}
 rules:
 - apiGroups:
   - jenkins.io
