+++ charts/lighthouse/templates/foghorn-sa.yaml
@@ -1,4 +1,4 @@
 kind: ServiceAccount
 apiVersion: v1
 metadata:
-  name: {{ template "status.name" . }}
+  name: {{ template "foghorn.name" . }}
